% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testsysfile.R
\name{testsysfile}
\alias{testsysfile}
\title{test system.file}
\usage{
testsysfile(sep = ",")
}
\arguments{
\item{sep}{character that separates records in the file}
}
\value{
contents of \code{inst/extdata/testfile.csv}.
}
\description{
test system.file in a package
}
\examples{
testsysfile()
}

